+++ client/package.json
@@ -9,7 +9,6 @@
   },
   "license": "ISC",
   "devDependencies": {
-    "@types/dat.gui": "^0.7.7",
     "@types/node": "^18.11.9",
     "@types/three": "^0.152.0",
     "peer": "^1.0.0",
@@ -22,7 +21,6 @@
     "webpack-merge": "^5.8.0"
   },
   "dependencies": {
-    "express": "^4.18.2",
     "peerjs": "^1.4.7"
   }
 }
